+++ create-loop/src/index.ts
@@ -174,6 +174,7 @@ const aptitudesPrompt = () => {
       { title: 'System', value: 'system'},
       { title: 'UI', value: 'ui' },
       { title: 'User', value: 'user' },
+      { title: 'Vault', value: 'vault' },
       { title: 'Window', value: 'window' },
     ],
     hint: 'Use your spacebar to select. You can select multiple!',