+++ loop-templates/templates/src/aptitudes/index.ts
@@ -7,6 +7,7 @@ import process from './process';
 import system from './system';
 import ui from './ui';
 import user from './user';
+import vault from './vault';
 import window from './window';
 import index from './index.ts.squirrelly';
 import { FileMap } from '@/types';
@@ -21,6 +22,7 @@ const fileMap: FileMap = {
   system: { fileName: 'system', aptitude: 'system'},
   ui: { fileName: 'ui', aptitude: 'any' },
   user: { fileName: 'user', aptitude: 'user' },
+  vault: { fileName: 'vault', aptitude: 'vault' },
   window: { fileName: 'window', aptitude: 'window' },
   index: { fileName: 'index.ts', aptitude: 'any' },
 };
@@ -35,6 +37,7 @@ export default {
   system,
   ui,
   user,
+  vault,
   window,
   index,
   fileMap,